Initialize chunky-operations-compose with a long int value.
This is a refinement to the type change made in a62f95180ef5c82869ec92d42cd82303039c5a24. Initializing with an int value for this benchmark leads to a similar issue I was trying to fix in that commit. In this case, we still end up with heterogenous arrays because the benchmark will store long int values to some of the array elements. By intializing with a long int, we can ensure the array only ever has one backing type.